SE_Proj/requirements.txt
mysql-connector-python
pycryptodome
PyPN
